fix(rmm): fix misra violations
fix misra-c2012-2.7, misra-c2012-10.7

+++ lib/common/include/utils_def.h
@@ -200,7 +200,7 @@
 	(((val) + UL(0)) << (regfield##_SHIFT))
 
 #define MASK(regfield) \
-	((~0UL >> (64UL - (regfield##_WIDTH))) << (regfield##_SHIFT))
+	((~0UL >> (unsigned int)(64UL - (regfield##_WIDTH))) << (regfield##_SHIFT))
 
 #define EXTRACT(regfield, reg) \
 	(((reg) & MASK(regfield)) >> (regfield##_SHIFT))
